Move the ParallelIteratorExt into the parallel_iterator_ext module
